Move default secret manager to stack
Thie removes a type test for stack type.
