+++ pkg/backend/httpstate/crypto_test.go
@@ -1,4 +1,4 @@
-package main
+package httpstate
 
 import (
 	"testing"
