+++ pkg/backend/mock.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/pulumi/pulumi/pkg/v3/operations"
 	"github.com/pulumi/pulumi/pkg/v3/resource/deploy"
+	"github.com/pulumi/pulumi/pkg/v3/secrets"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/apitype"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/diag"
 	sdkDisplay "github.com/pulumi/pulumi/sdk/v3/go/common/display"
@@ -355,8 +356,9 @@ type MockStack struct {
 	RenameF  func(ctx context.Context, newName tokens.QName) (StackReference, error)
 	GetLogsF func(ctx context.Context, cfg StackConfiguration,
 		query operations.LogQuery) ([]operations.LogEntry, error)
-	ExportDeploymentF func(ctx context.Context) (*apitype.UntypedDeployment, error)
-	ImportDeploymentF func(ctx context.Context, deployment *apitype.UntypedDeployment) error
+	ExportDeploymentF     func(ctx context.Context) (*apitype.UntypedDeployment, error)
+	ImportDeploymentF     func(ctx context.Context, deployment *apitype.UntypedDeployment) error
+	DefaultSecretManagerF func(configFile string) (secrets.Manager, error)
 }
 
 var _ Stack = (*MockStack)(nil)
@@ -481,6 +483,13 @@ func (ms *MockStack) ExportDeployment(ctx context.Context) (*apitype.UntypedDepl
 func (ms *MockStack) ImportDeployment(ctx context.Context, deployment *apitype.UntypedDeployment) error {
 	if ms.ImportDeploymentF != nil {
 		return ms.ImportDeploymentF(ctx, deployment)
+	}
+	panic("not implemented")
+}
+
+func (ms *MockStack) DefaultSecretManager(configFile string) (secrets.Manager, error) {
+	if ms.DefaultSecretManagerF != nil {
+		return ms.DefaultSecretManagerF(configFile)
 	}
 	panic("not implemented")
 }
